+++ src/main/java/com/fiaschetti/bankdemo/dto/DepositOperationDTO.java
@@ -1,5 +1,6 @@
 package com.fiaschetti.bankdemo.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class DepositOperationDTO extends OperationDTO {
@@ -8,7 +9,7 @@ public class DepositOperationDTO extends OperationDTO {
         super();
     }
 
-    public DepositOperationDTO(Date operationDate, String description, double amount) {
+    public DepositOperationDTO(Date operationDate, String description, BigDecimal amount) {
         super(operationDate, description, amount);
     }
